requirements.txt
Django>=4.1.3
djangorestframework
python-dotenv
psycopg2-binary
drf-spectacular
django-cors-headers
openai
PyPDF2
pytesseract
icalendar
requests